+++ backend/workoutapp/src/main/java/org/example/workoutapp/repository/ActivityRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import java.util.List;
 import java.util.Optional;
 
-public interface ActivityRepository extends JpaRepository<Activity, Long> {
+public interface ActivityRepository extends JpaRepository<Activity, Integer> {
 
     @Query(
             value="SELECT activity.*, COUNT(likes.like_id) as totalLikes " +
